server/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pycrypto==2.6.1
Werkzeug==0.15.2
